requirements.txt
beautifulsoup4>=4.12.2
fastapi>=0.96.1
feedparser>=6.0.10
html2text>=2020.1.16
langchain>=0.0.196
llama_index>=0.5.15
nltk>=3.8.1
openai>=0.27.2
protobuf>=4.23.2
pydantic>=1.10.7
pymongo>=4.3.3
python-dotenv>=1.0.0
qdrant_client>=1.2.0
requests>=2.28.2
tqdm>=4.65.0
trafilatura>=1.6.0
uvicorn>=0.22.0
